refactor(playground): extract platform/ helpers
Move framework-agnostic DOM utilities out of main.ts:

- platform/dom.ts: el() element factory
- platform/toast.ts: toast notification with auto-dismiss timer
- platform/hold-to-repeat.ts: button hold-to-repeat wiring

These have no app-state dependency and will be reused by multiple
features once the main.ts split continues.
