+++ playground/src/platform/dom.ts
@@ -1,0 +1,11 @@
+/** Build an HTML element with a className and optional text in one call. */
+export function el<K extends keyof HTMLElementTagNameMap>(
+  tag: K,
+  className: string,
+  text?: string,
+): HTMLElementTagNameMap[K] {
+  const node = document.createElement(tag);
+  node.className = className;
+  if (text !== undefined) node.textContent = text;
+  return node;
+}
