+++ playground/src/platform/index.ts
@@ -1,0 +1,3 @@
+export { el } from "./dom";
+export { toast } from "./toast";
+export { attachHoldToRepeat } from "./hold-to-repeat";
